+++ cmd/psweb/config.go
@@ -5,6 +5,8 @@ import (
 	"log"
 	"os"
 	"os/user"
+	"strconv"
+	"strings"
 )
 
 type Configuration struct {
@@ -16,37 +18,56 @@ type Configuration struct {
 	LiquidApi         string // for liquid tx links
 	NodeApi           string // for node links
 	MaxHistory        uint
-	ConfigFile        string
+	DataDir           string
+	ElementsPass      string
+	BitcoinSwaps      bool
+	Chain             string
+	LocalMempool      string
 }
 
-var Config Configuration
+var config Configuration
 
-func LoadConfig(configFile string) {
+func loadConfig(dataDir string) {
 
-	if configFile == "" {
+	if dataDir == "" {
 		// Get the current user's information
 		currentUser, err := user.Current()
 		if err != nil {
 			log.Fatalln(err)
 		}
-		// Default is /home/user/.peerswap/pswebconfig.json
-		configFile = currentUser.HomeDir + "/.peerswap/pswebconfig.json"
+		// Default is /home/user/.peerswap
+		dataDir = currentUser.HomeDir + "/.peerswap"
 	}
 
 	// load defaults first
-	Config.AllowSwapRequests = true
-	Config.RpcHost = "localhost:42069"
-	Config.ListenPort = "8088"
-	Config.ColorScheme = "dark" // dark or light
-	Config.NodeApi = "https://amboss.space/node"
-	Config.BitcoinApi = "https://mempool.space"
-	Config.LiquidApi = "https://liquid.network"
-	Config.MaxHistory = 10
-	Config.ConfigFile = configFile
+	config.AllowSwapRequests = true
+	config.RpcHost = "localhost:42069"
+	config.ListenPort = "1984"
+	config.ColorScheme = "dark" // dark or light
+	config.NodeApi = "https://amboss.space/node"
+	config.BitcoinApi = "https://mempool.space"
+	config.LiquidApi = "https://liquid.network"
+	config.MaxHistory = 10
+	config.DataDir = dataDir
+	config.ElementsPass = ""
+	config.BitcoinSwaps = true
+	config.Chain = "mainnet"
+	config.LocalMempool = ""
+
+	// environment values take priority
+	if os.Getenv("NETWORK") == "testnet" {
+		config.Chain = "testnet"
+		config.NodeApi = "https://mempool.space/testnet/lightning/node"
+		config.BitcoinApi = "https://mempool.space/testnet"
+		config.LiquidApi = "https://liquid.network/testnet"
+	}
+
+	configFile := dataDir + "/pswebconfig.json"
 
 	fileData, err := os.ReadFile(configFile)
 	if err != nil {
-		err = SaveConfig()
+		// save defaults in a newly created file
+		err = saveConfig()
 		if err != nil {
 			log.Println("Error creating config file.", err)
 		} else {
@@ -55,22 +76,103 @@ func LoadConfig(configFile string) {
 		return
 	}
 
-	err = json.Unmarshal(fileData, &Config)
+	err = json.Unmarshal(fileData, &config)
 	if err != nil {
 		log.Println("Error unmarshalling config file. Using defaults.")
 	}
 }
 
-func SaveConfig() error {
-	jsonData, err := json.MarshalIndent(Config, "", "  ")
+func saveConfig() error {
+	jsonData, err := json.MarshalIndent(config, "", "  ")
 	if err != nil {
 		return err
 	}
-	filename := Config.ConfigFile
+	filename := config.DataDir + "/pswebconfig.json"
 	err = os.WriteFile(filename, jsonData, 0644)
 	if err != nil {
 		return err
 	}
-
 	return nil
+}
+
+func savePeerSwapdConfig() {
+	// Get the current user's information
+	currentUser, err := user.Current()
+	if err != nil {
+		log.Fatalln(err)
+	}
+	filename := config.DataDir + "/peerswap.conf"
+	defaultLndDir := currentUser.HomeDir + "/.lnd"
+	//key, default, new value, env key
+	t := setPeerswapdVariable("host", "localhost:42069", config.RpcHost, "")
+	t += setPeerswapdVariable("rpchost", "localhost:42070", "", "")
+	t += setPeerswapdVariable("lnd.host", "localhost:10009", "", "LND_HOST")
+	t += setPeerswapdVariable("lnd.tlscertpath", defaultLndDir+"/tls.cert", "", "")
+	t += setPeerswapdVariable("lnd.macaroonpath", defaultLndDir+"/data/chain/bitcoin/"+config.Chain+"/admin.macaroon", "", "LND_MACAROONPATH")
+	t += setPeerswapdVariable("elementsd.rpcuser", "admin1", "", "ELEMENTSD_RPCUSER")
+	t += setPeerswapdVariable("elementsd.rpcpass", "123", config.ElementsPass, "")
+	t += setPeerswapdVariable("elementsd.rpchost", "http://127.0.0.1", "", "ELEMENTSD_RPCHOST")
+	t += setPeerswapdVariable("elementsd.rpcport", "18884", "", "ELEMENTSD_RPCPORT")
+	t += setPeerswapdVariable("elementsd.rpcwallet", "swaplnd", "", "ELEMENTSD_RPCWALLET")
+	t += setPeerswapdVariable("bitcoinswaps", "false", strconv.FormatBool(config.BitcoinSwaps), "")
+
+	data := []byte(t)
+
+	// Open the file in write-only mode, truncate if exists or create a new file
+	file, err := os.OpenFile(filename, os.O_WRONLY|os.O_TRUNC|os.O_CREATE, 0644)
+	if err != nil {
+		log.Println("Error opening file:", err)
+		return
+	}
+	defer file.Close()
+
+	// Write data to the file
+	_, err = file.Write(data)
+	if err != nil {
+		log.Println("Error writing to file:", err)
+		return
+	}
+}
+
+func setPeerswapdVariable(variableName, defaultValue, newValue, envKey string) string {
+	// returns variable=value string\n
+	// priority:
+	// 1. newValue
+	// 2. envValue
+	// 3. oldValue from peerswap.conf
+	// 4. defaultValue
+
+	v := defaultValue
+	if newValue != "" {
+		v = newValue
+	} else if envKey != "" && os.Getenv(envKey) != "" {
+		v = os.Getenv(envKey)
+	} else if s := readVariableFromPeerswapdConfig(variableName); s != "" {
+		v = s
+	}
+	return variableName + "=" + v + "\n"
+}
+
+func readVariableFromPeerswapdConfig(searchVariable string) string {
+	// Read the entire content of the file
+	filePath := config.DataDir + "/peerswap.conf"
+	content, err := os.ReadFile(filePath)
+	if err != nil {
+		return ""
+	}
+	// Convert the content to a string
+	fileContent := string(content)
+
+	if index := strings.Index(fileContent, searchVariable); index > 0 {
+		startIndex := index + len(searchVariable) + 1
+		value := ""
+		for _, char := range fileContent[startIndex:] {
+			if char == '\n' || char == '\r' {
+				break
+			}
+			value += string(char)
+		}
+		return value
+	}
+	return ""
 }